+++ modules/cdp_treasury/src/mock.rs
@@ -125,7 +125,7 @@ pub type Currencies = orml_currencies::Module<Runtime>;
 parameter_types! {
 	pub const GetStableCurrencyId: CurrencyId = AUSD;
 	pub const GetExchangeFee: (u32, u32) = (0, 100);
-	pub const TradingPathLimit: usize = 3;
+	pub const TradingPathLimit: u32 = 3;
 	pub EnabledTradingPairs : Vec<TradingPair> = vec![TradingPair::new(AUSD, BTC)];
 	pub const DEXModuleId: ModuleId = ModuleId(*b"aca/dexm");
 }
